[#206] Decompose PersonDetailView into child views with modal state enum
- Extract 5 child views: PersonHeroSection, PersonQuickActionsBar,
  PersonConversationContextCard, PersonTouchHistorySection,
  PersonSettingsSection
- Replace 12 boolean/optional @State modal vars with PersonDetailSheet
  and PersonDetailAlert enums (2 @State vars replace 12)
- Add PersonSettingsAction enum for child-to-parent communication
- Extract reusable datePickerSheet and undoBanner helpers
- PersonDetailView reduced from 1,168 to 426 lines (64% reduction)
- PersonDetailViewModel unchanged; all tests pass
